creating a new component to hold the data for the modal language
